refactor: 피드작성에 대한 이벤트에서 피드 세부정보 제거
